Improved the logic for generating a playlist for the user.
Now the auto generated playlist cover image is persisted and it is checked if the user is logged in with spotify account. There are small changes of the usings in the Automapper Profille as i forgot to stage it in the past (when i moved Spotify functionallity to separate assembly).